requirements.txt
streamlit
langchain
langchain_community
requests
psycopg2
GitPython